requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pyyaml>=6.0
reportlab>=3.6.0
markdown>=3.3.0
jinja2>=3.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
